added controller layer, cleaned code accordingly
