+++ controllers/README.md
@@ -1,0 +1,2 @@
+A controller's job is to validate the parameters and then interact to service to get data.
+It's responsible for generating correct response status code.
