+++ middlewares/index.js
@@ -3,7 +3,7 @@ const jwt = require('jsonwebtoken'); // used to create, sign, and verify tokens
 const env = process.env.NODE_ENV || 'development';
 const config = require('../config/config.js')[env];
 
-const utilityService = require('../services/utilityService');
+const { saveRequestLog } = require('../services/utility');
 const logger = require('../modules/logger');
 
 const statusCode = require('../constants/statusCode');
@@ -34,7 +34,7 @@ module.exports = {
      * so saving it into database
      */
     if (url.indexOf('/search/sellers') >= 0) {
-      utilityService.saveRequestLog({
+      saveRequestLog({
         ipAddress: ipAddress,
         url: url,
         body: JSON.stringify(body)
